Unify time to be in seconds in whole application. Fix runtime exceptions.

+++ src/core/ChangeableObject.java
@@ -3,5 +3,5 @@ package core;
 public interface ChangeableObject {
 
 
-    void update(long currentTimeMilliseconds);
+    void update(double currentTimeMilliseconds);
 }
