+++ src/visualComponents/StoreCheckout.java
@@ -230,7 +230,7 @@ public class StoreCheckout extends AnimatedAndObservable{
 
 
 	@Override
-	public void update(long currentTimeMilliseconds) {
+	public void update(double currentTimeMilliseconds) {
 
 	}
 }
